Refactor some codes related to lifecycle of the session's event stream

+++ broker/service/transaction_service.go
@@ -2,12 +2,12 @@ package service
 
 import (
 	"context"
-	"errors"
 	"github.com/golang/protobuf/proto"
 	"github.com/paust-team/shapleq/broker/internals"
 	"github.com/paust-team/shapleq/broker/service/rpc"
 	"github.com/paust-team/shapleq/broker/storage"
 	"github.com/paust-team/shapleq/message"
+	"github.com/paust-team/shapleq/pqerror"
 	shapleqproto "github.com/paust-team/shapleq/proto"
 	"github.com/paust-team/shapleq/zookeeper"
 	"runtime"
@@ -49,24 +49,12 @@ func (s *TransactionService) HandleEventStreams(brokerCtx context.Context, event
 			select {
 			case <-brokerCtx.Done():
 				return
-			case eventStream := <-eventStreamCh:
-				wg.Add(1)
-				go func() {
-					defer wg.Done()
-					for {
-						select {
-						case msg := <-eventStream.MsgCh:
-							if msg == nil {
-								return
-							}
-							if err := s.handleMsg(msg, eventStream.Session); err != nil {
-								errCh <- err
-							}
-						default:
-						}
-						runtime.Gosched()
-					}
-				}()
+			case eventStream, ok := <-eventStreamCh:
+				if ok {
+					wg.Add(1)
+					go s.handleEventStream(eventStream, errCh, &wg)
+				}
+
 			default:
 			}
 			runtime.Gosched()
@@ -76,37 +64,57 @@ func (s *TransactionService) HandleEventStreams(brokerCtx context.Context, event
 	return errCh
 }
 
-func (s *TransactionService) handleMsg(msg *message.QMessage, session *internals.Session) error {
+func (s *TransactionService) handleEventStream(eventStream internals.EventStream, sessionErrCh chan error, wg *sync.WaitGroup) {
 
-	var resMsg proto.Message
+	defer wg.Done()
+	readCh := eventStream.ReadMsgCh
+	writeCh := eventStream.WriteMsgCh
 
-	if reqMsg, err := msg.UnpackAs(&shapleqproto.CreateTopicRequest{}); err == nil {
-		resMsg = s.rpcService.CreateTopic(reqMsg.(*shapleqproto.CreateTopicRequest))
+	defer close(writeCh)
 
-	} else if reqMsg, err := msg.UnpackAs(&shapleqproto.DeleteTopicRequest{}); err == nil {
-		resMsg = s.rpcService.DeleteTopic(reqMsg.(*shapleqproto.DeleteTopicRequest))
+	for {
+		select {
+		case msg, ok := <-readCh:
+			if ok {
+				var resMsg proto.Message
 
-	} else if reqMsg, err := msg.UnpackAs(&shapleqproto.ListTopicRequest{}); err == nil {
-		resMsg = s.rpcService.ListTopic(reqMsg.(*shapleqproto.ListTopicRequest))
+				if reqMsg, err := msg.UnpackAs(&shapleqproto.CreateTopicRequest{}); err == nil {
+					resMsg = s.rpcService.CreateTopic(reqMsg.(*shapleqproto.CreateTopicRequest))
 
-	} else if reqMsg, err := msg.UnpackAs(&shapleqproto.DescribeTopicRequest{}); err == nil {
-		resMsg = s.rpcService.DescribeTopic(reqMsg.(*shapleqproto.DescribeTopicRequest))
+				} else if reqMsg, err := msg.UnpackAs(&shapleqproto.DeleteTopicRequest{}); err == nil {
+					resMsg = s.rpcService.DeleteTopic(reqMsg.(*shapleqproto.DeleteTopicRequest))
 
-	} else if reqMsg, err := msg.UnpackAs(&shapleqproto.Ping{}); err == nil {
-		resMsg = s.rpcService.Heartbeat(reqMsg.(*shapleqproto.Ping))
+				} else if reqMsg, err := msg.UnpackAs(&shapleqproto.ListTopicRequest{}); err == nil {
+					resMsg = s.rpcService.ListTopic(reqMsg.(*shapleqproto.ListTopicRequest))
 
-	} else if reqMsg, err := msg.UnpackAs(&shapleqproto.DiscoverBrokerRequest{}); err == nil {
-		resMsg = s.rpcService.DiscoverBroker(reqMsg.(*shapleqproto.DiscoverBrokerRequest))
-	} else {
-		return errors.New("invalid message to handle")
+				} else if reqMsg, err := msg.UnpackAs(&shapleqproto.DescribeTopicRequest{}); err == nil {
+					resMsg = s.rpcService.DescribeTopic(reqMsg.(*shapleqproto.DescribeTopicRequest))
+
+				} else if reqMsg, err := msg.UnpackAs(&shapleqproto.Ping{}); err == nil {
+					resMsg = s.rpcService.Heartbeat(reqMsg.(*shapleqproto.Ping))
+
+				} else if reqMsg, err := msg.UnpackAs(&shapleqproto.DiscoverBrokerRequest{}); err == nil {
+					resMsg = s.rpcService.DiscoverBroker(reqMsg.(*shapleqproto.DiscoverBrokerRequest))
+				} else {
+					sessionErrCh <- err
+				}
+
+				qMsg, err := message.NewQMessageFromMsg(message.TRANSACTION, resMsg)
+				if err != nil {
+					sessionErrCh <- internals.TransactionalError{
+						PQError:       err.(pqerror.PQError),
+						Session:       eventStream.Session,
+						CancelSession: eventStream.CancelSession,
+					}
+				} else {
+					writeCh <- qMsg
+				}
+			} else {
+				return
+			}
+		default:
+
+		}
+		runtime.Gosched()
 	}
-
-	qMsg, err := message.NewQMessageFromMsg(message.TRANSACTION, resMsg)
-	if err != nil {
-		return err
-	}
-	if err := session.Write(qMsg); err != nil {
-		return err
-	}
-	return nil
 }
